+++ biopipen/ns/vcf.py
@@ -18,7 +18,6 @@ class VcfLiftOver(Proc):
         tmpdir: Directory for temporary storage of working files
         args: Other CLI arguments for `gatk LiftoverVcf`
     """
-
     input = "invcf:file"
     output = "outvcf:file:{{in.invcf | basename}}"
     envs = {
@@ -61,7 +60,6 @@ class VcfFilter(Proc):
         helper: Some helper code for the filters
         keep: Keep the variants not passing the filters?
     """  # noqa: E501
-
     input = "invcf:file"
     output = "outfile:file:{{in.invcf | basename}}"
     lang = config.lang.python
@@ -87,7 +85,6 @@ class VcfIndex(Proc):
     Envs:
         tabix: Path to tabix
     """
-
     input = "infile:file"
     output = """
         {%- if in.infile.endswith(".gz") %}
@@ -120,11 +117,9 @@ class Vcf2Bed(Proc):
         outbase: The coordinate base of the base file
 
     Requires:
-        - name: cyvcf2
-          check: |
-            {{proc.lang}} -c "import cyvcf2"
+        cyvcf2:
+            - check: {{proc.lang}} -c "import cyvcf2"
     """
-
     input = "infile:file"
     output = "outfile:file:{{in.infile | stem0}}.bed"
     lang = config.lang.python
@@ -147,10 +142,10 @@ class VcfDownSample(Proc):
             If `n > 1`, it is the number.
             If `n <= 1`, it is the fraction.
     """
-
     input = "infile:file"
     output = "outfile:file:{{in.infile | basename}}"
     envs = {"n": 0}
+    lang = config.lang.bash
     script = "file://../scripts/vcf/VcfDownSample.sh"
 
 
@@ -290,19 +285,50 @@ class VcfFix(Proc):
         helpers: raw code the provide some helpers for the fixes
             The code will automatically dedented if given as a string. A list
             of strings is also supported and will be joined with newlines.
-
-    Requires:
-        - name: biopipen
-          check: |
-            {{proc.lang}} -c "import biopipen"
-
     """
-
     input = "infile:file"
     output = "outfile:file:{{in.infile | basename}}"
     lang = config.lang.python
     envs = {"fixes": [], "helpers": ""}
     script = "file://../scripts/vcf/VcfFix.py"
+
+
+class VcfAnno(Proc):
+    """Annotate a VCF file using vcfanno
+
+    https://github.com/brentp/vcfanno
+
+    Input:
+        infile: The input VCF file
+        conffile: The configuration file for vcfanno or configuration dict
+            itself
+
+    Output:
+        outfile: The output VCF file
+
+    Envs:
+        vcfanno: Path to vcfanno
+        ncores: Number of cores to use
+        conffile: configuration file for vcfanno or configuration dict itself
+            This is ignored when `conffile` is given as input
+        args: Additional arguments to pass to vcfanno
+
+    Requires:
+        - name: vcfanno
+          check: |
+            {{proc.envs.vcfanno}} --help
+    """
+
+    input = "infile:file, conffile"
+    output = "outfile:file:{{in.infile | stem0}}.{{envs.tool}}.vcf"
+    lang = config.lang.python
+    envs = {
+        "vcfanno": config.exe.vcfanno,
+        "ncores": config.misc.ncores,
+        "conffile": {},
+        "args": {"permissive-overlap": True},
+    }
+    script = "file://../scripts/vcf/VcfAnno.py"
 
 
 class TruvariBench(Proc):
@@ -323,11 +349,9 @@ class TruvariBench(Proc):
         `<other>`: Ohter `truvari bench` arguments
 
     Requires:
-        - name: truvari
-          check: |
-            {{proc.envs.truvari}} version
+        truvari:
+            - check: {{proc.envs.truvari}} version
     """
-
     input = "compvcf:file, basevcf:file"
     output = "outdir:dir:{{in.compvcf | stem0 | append: '.truvari_bench'}}"
     envs = {
@@ -340,6 +364,7 @@ class TruvariBench(Proc):
         "typeignore": False,
         "multimatch": False,
     }
+    lang = config.lang.bash
     script = "file://../scripts/vcf/TruvariBench.sh"
 
 
@@ -363,21 +388,15 @@ class TruvariBenchSummary(Proc):
         devpars: The parameters to use for the plots.
 
     Requires:
-        - name: r-ggprism
-          check: |
-            {{proc.lang}} -e "library(ggprism)"
-        - name: r-rjson
-          check: |
-            {{proc.lang}} -e "library(rjson)"
-        - name: r-dplyr
-          check: |
-            {{proc.lang}} -e "library(dplyr)"
-        - name: r-ggplot2
-          check: |
-            {{proc.lang}} -e "library(ggplot2)"
-
+        r-ggprism:
+            - check: {{proc.lang}} -e "library(ggprism)"
+        r-rjson:
+            - check: {{proc.lang}} -e "library(rjson)"
+        r-dplyr:
+            - check: {{proc.lang}} -e "library(dplyr)"
+        r-ggplot2:
+            - check: {{proc.lang}} -e "library(ggplot2)"
     """
-
     input = "indirs:files"
     input_data = lambda ch: [list(ch.iloc[:, 0])]
     output = "outdir:dir:truvari_bench.summary"
@@ -411,7 +430,6 @@ class TruvariConsistency(Proc):
             annotations will be added as row annotations.
             Other options see also `biopipen.ns.plot.Heatmap`.
     """
-
     input = "vcfs:files"
     output = (
         "outdir:dir:"