+++ biopipen/reports/bam/ControlFREEC.svelte
@@ -1,6 +1,6 @@
 {% from "utils/misc.liq" import report_jobs, table_of_images -%}
 <script>
-    import { Image, DataTable } from "@@";
+    import { Image, DataTable } from "$lib";
 </script>
 
 {%- macro report_job(job, h=1) -%}