+++ biopipen/reports/scrna_metabolic_landscape/MetabolicPathwayHeterogeneity.svelte
@@ -1,7 +1,7 @@
 {% from "utils/misc.liq" import report_jobs, table_of_images -%}
 
 <script>
-    import { Image } from "@@";
+    import { Image } from "$lib";
 </script>
 
 {%- macro report_job(job, h=2) -%}