+++ biopipen/reports/tcr/SampleDiversity.svelte
@@ -1,7 +1,7 @@
 {% from "utils/misc.liq" import report_jobs, table_of_images -%}
 <script>
-    import { Image } from "@@";
-    import { Tile } from "carbon-components-svelte";
+    import { Image } from "$lib";
+    import { Tile } from "$ccs";
 </script>