+++ biopipen/reports/tcr/TCRClusteringStats.svelte
@@ -1,7 +1,7 @@
 {% from "utils/misc.liq" import report_jobs -%}
 <script>
-    import { Image, DataTable } from "@@";
-    import { Tabs, Tab, TabContent } from "carbon-components-svelte";
+    import { Image, DataTable } from "$lib";
+    import { Tabs, Tab, TabContent } from "$ccs";
 </script>
 
 {%- macro report_job(job, h=1) -%}